CHROMIUM: gobi: make tlv_get return an int so -errno works.
tlv_get returns -ENOMEM or -ENOMSG, but it was declared as returning u16,
so those errors got squashed.

BUG=None
TEST=None (nobody currently checks for -errno anyway, just specific values)


Reviewed-on: http://gerrit.chromium.org/gerrit/4501
Tested-by: ttuttle <ttuttle@chromium.org>

+++ drivers/net/usb/gobi/qmi.c
@@ -215,7 +215,7 @@ int qmux_fill(u16 cid, struct buffer *buf)
 	return 0;
 }
 
-static u16 tlv_get(void *msg, u16 msgsize, u8 type, void *buf, u16 bufsize)
+static int tlv_get(void *msg, u16 msgsize, u8 type, void *buf, u16 bufsize)
 {
 	u16 pos;
 	u16 msize = 0;
